Adjust ToolShedRepositoryCache tests to new tool_path / repository_path attributes
